add graph values base class
